Added float to string function to logging
To support Logging of floating point values, I had to add a cludge to
the logging functions.
As it turns out, Arduino's "sprintf" function does not support floating
point values. To get around this, I added a function to convert the
floating point values to strings. This function would need to be called
before logging a float.