Re build (#1)
* Re-Build using C Implementation

This is the port of the C Implementaiton Re-build into arduino.
It compiles, but full verification has not been done.

* Foud Arduino Bug

As it turns out, Arduino's "sprintf" does not support float types. i.e.
%f.
I don't have a good workaround yet, but I'm working on it...

* Added float to string function to logging

To support Logging of floating point values, I had to add a cludge to
the logging functions.
As it turns out, Arduino's "sprintf" function does not support floating
point values. To get around this, I added a function to convert the
floating point values to strings. This function would need to be called
before logging a float.

* Finished fixxing the Logging function

I don't like it, but it doesn't seem like I've a choice...

* Finalized changes from C emulation update

Nearing end of rebuild.
Last steps will be to comb through code and make sure I've propperly
incorporated the control struct.

* Finished implementing gapa code

about to test (non-official, from home)